fix test

+++ symengine/printers/latex.cpp
@@ -55,7 +55,7 @@ void LatexPrinter::bvisit(const Symbol &x)
         int count = 0;
         std::string prev = str_;
         str_ = "";
-        for (int i = 0; i < prev.size(); i++) {
+        for (size_t i = 0; i < prev.size(); i++) {
             const char &c = prev[i];
             if (c == '_' and prev.size() > i + 2) {
                 str_ += "_{";
